Add files via upload
The following is the capstone 2 project assignment given by #Purwadhika and my mentor Ka Median. In this project, I analyzed the Supermarket Customer data set. I will analyze the characteristics of customers who shop frequently and participate in campaigns. Apart from that, I also analyze the products that are frequently purchased and the campaigns that are most followed by customers.